fix can not logout when session timeout

+++ src/main/java/com/youwei/zjb/user/UserService.java
@@ -300,8 +300,10 @@ public class UserService {
 	public ModelAndView logout(PC pc){
 		ModelAndView mv = new ModelAndView();
 		//httpsession timeout & remove sessionid from db
-		dao.execute("delete from UserSession where userid=? and sessionId=?", ThreadSession.getUser().id, ThreadSession.getHttpSession().getId());
-		ThreadSession.getHttpSession().invalidate();
+		if(ThreadSession.getUser()!=null){
+			dao.execute("delete from UserSession where userid=? and sessionId=?", ThreadSession.getUser().id, ThreadSession.getHttpSession().getId());
+			ThreadSession.getHttpSession().invalidate();
+		}
 		return mv;
 	}
